Refactor LayoutHeader/Nav/Popup hook * move useNavPopups hook to Popup/util.tsx * split useNavPopups into multiple functions
